monitor/requirements.txt
requests>=2.31.0
pyyaml>=6.0
pytz>=2024.1
